fix a bug in query inference
The use of `some` was causing `infer` to only find a single match for
any given key, when in reality keys can be arbitrarily nested in
multiple places in a pull spec. The fix replaces the internal
`get-binding` fn with the new `get-bindings` helper, which mapcats
all pairs together, and partitioning the result.